move buildName to be a test helper
